+++ pybamm/experiment/step/__init__.py
@@ -1,3 +1,3 @@
 from .steps import *
-from .steps import _Step
+from .base_step import BaseStep, BaseStepExplicit, BaseStepImplicit
 from .step_termination import *