+++ pybamm/simulation.py
@@ -103,7 +103,7 @@ class Simulation:
                         }
                     )
         else:
-            if isinstance(experiment, (str, pybamm.step._Step)):
+            if isinstance(experiment, (str, pybamm.step.BaseStep)):
                 experiment = pybamm.Experiment([experiment])
             elif isinstance(experiment, list):
                 experiment = pybamm.Experiment(experiment)
@@ -131,8 +131,8 @@ class Simulation:
         self._model_with_set_params = None
         self._built_model = None
         self._built_initial_soc = None
-        self.op_conds_to_built_models = None
-        self.op_conds_to_built_solvers = None
+        self.steps_to_built_models = None
+        self.steps_to_built_solvers = None
         self._mesh = None
         self._disc = None
         self._solution = None
@@ -178,149 +178,41 @@ class Simulation:
 
     def set_up_and_parameterise_experiment(self):
         """
-        Set up a simulation to run with an experiment. This creates a dictionary of
-        inputs (current/voltage/power, running time, stopping condition) for each
-        operating condition in the experiment. The model will then be solved by
-        integrating the model successively with each group of inputs, one group at a
-        time.
-        This needs to be done here and not in the Experiment class because the nominal
-        cell capacity (from the parameters) is used to convert C-rate to current.
-        """
-        # Update experiment using capacity
-        capacity = self._parameter_values["Nominal cell capacity [A.h]"]
-        for op_conds in self.experiment.operating_conditions_steps:
-            if op_conds.type == "C-rate":
-                op_conds.type = "current"
-                op_conds.value = op_conds.value * capacity
-
-            # Add time to the experiment times
-            dt = op_conds.duration
-            if dt is None:
-                if op_conds.type == "current":
-                    # Current control: max simulation time: 3h / C-rate
-                    Crate = op_conds.value / capacity
-                    dt = 3 / abs(Crate) * 3600  # seconds
-                else:
-                    # max simulation time: 1 day
-                    dt = 24 * 3600  # seconds
-                op_conds.duration = dt
-
-        # Set up model for experiment
-        self.set_up_and_parameterise_model_for_experiment()
-
-    def set_up_and_parameterise_model_for_experiment(self):
-        """
-        Set up self._model to be able to run the experiment (new version).
-        In this version, a new model is created for each step.
+        Create and parameterise the models for each step in the experiment.
 
         This increases set-up time since several models to be processed, but
         reduces simulation time since the model formulation is efficient.
         """
+        parameter_values = self._parameter_values.copy()
+        # Set the initial temperature to be the temperature of the first step
+        # We can set this globally for all steps since any subsequent steps will either
+        # start at the temperature at the end of the previous step (if non-isothermal
+        # model), or will use the "Ambient temperature" input (if isothermal model).
+        # In either case, the initial temperature is not used for any steps except
+        # the first.
+        init_temp = self.experiment.steps[0].temperature
+        if init_temp is not None:
+            parameter_values["Initial temperature [K]"] = init_temp
+
+        # Process each step
         self.experiment_unique_steps_to_model = {}
-        for op_number, op in enumerate(self.experiment.unique_steps):
-            new_model = self._model.new_copy()
-            new_parameter_values = self._parameter_values.copy()
-
-            if op.type != "current":
-                # Voltage or power control
-                # Create a new model where the current density is now a variable
-                # To do so, we replace all instances of the current density in the
-                # model with a current density variable, which is obtained from the
-                # FunctionControl submodel
-                # check which kind of external circuit model we need (differential
-                # or algebraic)
-                if op.type == "voltage":
-                    submodel_class = pybamm.external_circuit.VoltageFunctionControl
-                elif op.type == "power":
-                    submodel_class = pybamm.external_circuit.PowerFunctionControl
-
-                # Build the new submodel and update the model with it
-                submodel = submodel_class(new_model.param, new_model.options)
-                variables = new_model.variables
-                submodel.variables = submodel.get_fundamental_variables()
-                variables.update(submodel.variables)
-                submodel.variables.update(submodel.get_coupled_variables(variables))
-                variables.update(submodel.variables)
-                submodel.set_rhs(variables)
-                submodel.set_algebraic(variables)
-                submodel.set_initial_conditions(variables)
-                new_model.rhs.update(submodel.rhs)
-                new_model.algebraic.update(submodel.algebraic)
-                new_model.initial_conditions.update(submodel.initial_conditions)
-
-                # Set the "current function" to be the variable defined in the submodel
-                new_parameter_values["Current function [A]"] = submodel.variables[
-                    "Current [A]"
-                ]
-            self.update_new_model_events(new_model, op)
-            # Update parameter values
-            self._original_temperature = new_parameter_values["Ambient temperature [K]"]
-            experiment_parameter_values = self.get_experiment_parameter_values(
-                op, op_number
-            )
-            new_parameter_values.update(
-                experiment_parameter_values, check_already_exists=False
-            )
-            parameterised_model = new_parameter_values.process_model(
-                new_model, inplace=False
-            )
-            self.experiment_unique_steps_to_model[op.basic_repr()] = parameterised_model
-
-        # Set up rest model if experiment has start times
-        if self.experiment.initial_start_time:
-            new_model = self._model.new_copy()
-            # Update parameter values
-            new_parameter_values = self._parameter_values.copy()
-            self._original_temperature = new_parameter_values["Ambient temperature [K]"]
-            new_parameter_values.update(
-                {"Current function [A]": 0, "Ambient temperature [K]": "[input]"},
-                check_already_exists=False,
-            )
-            parameterised_model = new_parameter_values.process_model(
-                new_model, inplace=False
-            )
-            self.experiment_unique_steps_to_model["Rest for padding"] = (
+        for step in self.experiment.unique_steps:
+            parameterised_model = step.process_model(self._model, parameter_values)
+            self.experiment_unique_steps_to_model[step.basic_repr()] = (
                 parameterised_model
             )
 
-    def update_new_model_events(self, new_model, op):
-        for term in op.termination:
-            event = term.get_event(new_model.variables, op.value)
-            if event is not None:
-                new_model.events.append(event)
-
-        # Keep the min and max voltages as safeguards but add some tolerances
-        # so that they are not triggered before the voltage limits in the
-        # experiment
-        for i, event in enumerate(new_model.events):
-            if event.name in ["Minimum voltage [V]", "Maximum voltage [V]"]:
-                new_model.events[i] = pybamm.Event(
-                    event.name, event.expression + 1, event.event_type
-                )
-
-    def get_experiment_parameter_values(self, op, op_number):
-        experiment_parameter_values = {
-            f"{op.type.capitalize()} function {op.unit}": op.value
-        }
-
-        if op.temperature is not None:
-            ambient_temperature = op.temperature
-            experiment_parameter_values.update(
-                {"Ambient temperature [K]": ambient_temperature}
+        # Set up rest model if experiment has start times
+        if self.experiment.initial_start_time:
+            # duration doesn't matter, we just need the model
+            rest_step = pybamm.step.rest(duration=1)
+            # Change ambient temperature to be an input, which will be changed at
+            # solve time
+            parameter_values["Ambient temperature [K]"] = "[input]"
+            parameterised_model = rest_step.process_model(self._model, parameter_values)
+            self.experiment_unique_steps_to_model["Rest for padding"] = (
+                parameterised_model
             )
-
-            # If at the first operation, then the intial temperature
-            # should be the ambient temperature.
-            if op_number == 0:
-                experiment_parameter_values.update(
-                    {"Initial temperature [K]": ambient_temperature}
-                )
-        else:
-            experiment_parameter_values.update(
-                {"Ambient temperature [K]": self._original_temperature}
-            )
-
-        return experiment_parameter_values
 
     def set_parameters(self):
         """
@@ -336,13 +228,13 @@ class Simulation:
         self._parameter_values.process_geometry(self._geometry)
         self._model = self._model_with_set_params
 
-    def set_initial_soc(self, initial_soc):
+    def set_initial_soc(self, initial_soc, inputs=None):
         if self._built_initial_soc != initial_soc:
             # reset
             self._model_with_set_params = None
             self._built_model = None
-            self.op_conds_to_built_models = None
-            self.op_conds_to_built_solvers = None
+            self.steps_to_built_models = None
+            self.steps_to_built_solvers = None
 
         options = self.model.options
         param = self._model.param
@@ -355,20 +247,28 @@ class Simulation:
         elif options["working electrode"] == "positive":
             self._parameter_values = (
                 self._unprocessed_parameter_values.set_initial_stoichiometry_half_cell(
-                    initial_soc, param=param, inplace=False, options=options
+                    initial_soc,
+                    param=param,
+                    inplace=False,
+                    options=options,
+                    inputs=inputs,
                 )
             )
         else:
             self._parameter_values = (
                 self._unprocessed_parameter_values.set_initial_stoichiometries(
-                    initial_soc, param=param, inplace=False, options=options
+                    initial_soc,
+                    param=param,
+                    inplace=False,
+                    options=options,
+                    inputs=inputs,
                 )
             )
 
         # Save solved initial SOC in case we need to re-build the model
         self._built_initial_soc = initial_soc
 
-    def build(self, check_model=True, initial_soc=None):
+    def build(self, check_model=True, initial_soc=None, inputs=None):
         """
         A method to build the model into a system of matrices and vectors suitable for
         performing numerical computations. If the model has already been built or
@@ -387,7 +287,7 @@ class Simulation:
             set.
         """
         if initial_soc is not None:
-            self.set_initial_soc(initial_soc)
+            self.set_initial_soc(initial_soc, inputs=inputs)
 
         if self.built_model:
             return
@@ -404,15 +304,15 @@ class Simulation:
             # rebuilt model so clear solver setup
             self._solver._model_set_up = {}
 
-    def build_for_experiment(self, check_model=True, initial_soc=None):
+    def build_for_experiment(self, check_model=True, initial_soc=None, inputs=None):
         """
         Similar to :meth:`Simulation.build`, but for the case of simulating an
         experiment, where there may be several models and solvers to build.
         """
         if initial_soc is not None:
-            self.set_initial_soc(initial_soc)
+            self.set_initial_soc(initial_soc, inputs)
 
-        if self.op_conds_to_built_models:
+        if self.steps_to_built_models:
             return
         else:
             self.set_up_and_parameterise_experiment()
@@ -424,10 +324,10 @@ class Simulation:
             self._mesh = pybamm.Mesh(self._geometry, self._submesh_types, self._var_pts)
             self._disc = pybamm.Discretisation(self._mesh, self._spatial_methods)
             # Process all the different models
-            self.op_conds_to_built_models = {}
-            self.op_conds_to_built_solvers = {}
+            self.steps_to_built_models = {}
+            self.steps_to_built_solvers = {}
             for (
-                op_cond,
+                step,
                 model_with_set_params,
             ) in self.experiment_unique_steps_to_model.items():
                 # It's ok to modify the model with set parameters in place as it's
@@ -436,8 +336,8 @@ class Simulation:
                     model_with_set_params, inplace=True, check_model=check_model
                 )
                 solver = self._solver.copy()
-                self.op_conds_to_built_solvers[op_cond] = solver
-                self.op_conds_to_built_models[op_cond] = built_model
+                self.steps_to_built_solvers[step] = solver
+                self.steps_to_built_models[step] = built_model
 
     def solve(
         self,
@@ -450,6 +350,7 @@ class Simulation:
         initial_soc=None,
         callbacks=None,
         showprogress=False,
+        inputs=None,
         **kwargs,
     ):
         """
@@ -512,8 +413,10 @@ class Simulation:
         callbacks = pybamm.callbacks.setup_callbacks(callbacks)
         logs = {}
 
+        inputs = inputs or {}
+
         if self.operating_mode in ["without experiment", "drive cycle"]:
-            self.build(check_model=check_model, initial_soc=initial_soc)
+            self.build(check_model=check_model, initial_soc=initial_soc, inputs=inputs)
             if save_at_cycles is not None:
                 raise ValueError(
                     "'save_at_cycles' option can only be used if simulating an "
@@ -584,11 +487,15 @@ class Simulation:
                             stacklevel=2,
                         )
 
-            self._solution = solver.solve(self.built_model, t_eval, **kwargs)
+            self._solution = solver.solve(
+                self.built_model, t_eval, inputs=inputs, **kwargs
+            )
 
         elif self.operating_mode == "with experiment":
             callbacks.on_experiment_start(logs)
-            self.build_for_experiment(check_model=check_model, initial_soc=initial_soc)
+            self.build_for_experiment(
+                check_model=check_model, initial_soc=initial_soc, inputs=inputs
+            )
             if t_eval is not None:
                 pybamm.logger.warning(
                     "Ignoring t_eval as solution times are specified by the experiment"
@@ -597,7 +504,7 @@ class Simulation:
             # inputs without having to build the simulation again
             self._solution = starting_solution
             # Step through all experimental conditions
-            user_inputs = kwargs.get("inputs", {})
+            user_inputs = inputs
             timer = pybamm.Timer()
 
             # Set up eSOH solver (for summary variables)
@@ -616,6 +523,7 @@ class Simulation:
                     [starting_solution],
                     esoh_solver=esoh_solver,
                     save_this_cycle=True,
+                    inputs=user_inputs,
                 )
                 starting_solution_cycles = [cycle_solution]
                 starting_solution_summary_variables = [cycle_sum_vars]
@@ -660,10 +568,10 @@ class Simulation:
             # Add initial padding rest if current time is earlier than first start time
             # This could be the case when using a starting solution
             if starting_solution is not None:
-                op_conds = self.experiment.operating_conditions_steps[0]
-                if op_conds.start_time is not None:
+                step = self.experiment.steps[0]
+                if step.start_time is not None:
                     rest_time = (
-                        op_conds.start_time
+                        step.start_time
                         - (
                             initial_start_time
                             + timedelta(seconds=float(current_solution.t[-1]))
@@ -673,10 +581,11 @@ class Simulation:
                         # logs["step operating conditions"] = "Initial rest for padding"
                         # callbacks.on_step_start(logs)
 
-                        kwargs["inputs"] = {
+                        inputs = {
                             **user_inputs,
                             "Ambient temperature [K]": (
-                                op_conds.temperature or self._original_temperature
+                                step.temperature
+                                or self._parameter_values["Ambient temperature [K]"]
                             ),
                             "start time": current_solution.t[-1],
                         }
@@ -684,7 +593,7 @@ class Simulation:
                         step_solution = current_solution.cycles[-1].steps[-1]
 
                         step_solution_with_rest = self.run_padding_rest(
-                            kwargs, rest_time, step_solution
+                            kwargs, rest_time, step_solution, inputs
                         )
                         steps[-1] = step_solution + step_solution_with_rest
 
@@ -745,23 +654,15 @@ class Simulation:
                 for step_num in range(1, cycle_length + 1):
                     # Use 1-indexing for printing cycle number as it is more
                     # human-intuitive
-                    op_conds = self.experiment.operating_conditions_steps[idx]
-
-                    # Hacky patch to allow correct processing of end_time and next_starting time
-                    # For efficiency purposes, op_conds treats identical steps as the same object
-                    # regardless of the initial time. Should be refactored as part of #3176
-                    op_conds_unproc = (
-                        self.experiment.operating_conditions_steps_unprocessed[idx]
-                    )
-
+                    step = self.experiment.steps[idx]
                     start_time = current_solution.t[-1]
 
                     # If step has an end time, dt must take that into account
-                    if getattr(op_conds_unproc, "end_time", None):
+                    if step.end_time is not None:
                         dt = min(
-                            op_conds.duration,
+                            step.duration,
                             (
-                                op_conds_unproc.end_time
+                                step.end_time
                                 - (
                                     initial_start_time
                                     + timedelta(seconds=float(start_time))
@@ -769,21 +670,21 @@ class Simulation:
                             ).total_seconds(),
                         )
                     else:
-                        dt = op_conds.duration
-                    op_conds_str = str(op_conds)
-                    model = self.op_conds_to_built_models[op_conds.basic_repr()]
-                    solver = self.op_conds_to_built_solvers[op_conds.basic_repr()]
+                        dt = step.duration
+                    step_str = str(step)
+                    model = self.steps_to_built_models[step.basic_repr()]
+                    solver = self.steps_to_built_solvers[step.basic_repr()]
 
                     logs["step number"] = (step_num, cycle_length)
-                    logs["step operating conditions"] = op_conds_str
+                    logs["step operating conditions"] = step_str
                     callbacks.on_step_start(logs)
 
-                    kwargs["inputs"] = {
+                    inputs = {
                         **user_inputs,
                         "start time": start_time,
                     }
                     # Make sure we take at least 2 timesteps
-                    npts = max(int(round(dt / op_conds.period)) + 1, 2)
+                    npts = max(int(round(dt / step.period)) + 1, 2)
                     try:
                         step_solution = solver.step(
                             current_solution,
@@ -791,6 +692,7 @@ class Simulation:
                             dt,
                             t_eval=np.linspace(0, dt, npts),
                             save=False,
+                            inputs=inputs,
                             **kwargs,
                         )
                     except pybamm.SolverError as error:
@@ -814,9 +716,9 @@ class Simulation:
                     step_termination = step_solution.termination
 
                     # Add a padding rest step if necessary
-                    if getattr(op_conds_unproc, "next_start_time", None) is not None:
+                    if step.next_start_time is not None:
                         rest_time = (
-                            op_conds_unproc.next_start_time
+                            step.next_start_time
                             - (
                                 initial_start_time
                                 + timedelta(seconds=float(step_solution.t[-1]))
@@ -827,16 +729,17 @@ class Simulation:
                             logs["step operating conditions"] = "Rest for padding"
                             callbacks.on_step_start(logs)
 
-                            kwargs["inputs"] = {
+                            inputs = {
                                 **user_inputs,
                                 "Ambient temperature [K]": (
-                                    op_conds.temperature or self._original_temperature
+                                    step.temperature
+                                    or self._parameter_values["Ambient temperature [K]"]
                                 ),
                                 "start time": step_solution.t[-1],
                             }
 
                             step_solution_with_rest = self.run_padding_rest(
-                                kwargs, rest_time, step_solution
+                                kwargs, rest_time, step_solution, inputs=inputs
                             )
                             step_solution += step_solution_with_rest
 
@@ -883,7 +786,7 @@ class Simulation:
                     if all(isinstance(step, pybamm.EmptySolution) for step in steps):
                         if len(steps) == 1:
                             raise pybamm.SolverError(
-                                f"Step '{op_conds_str}' is infeasible "
+                                f"Step '{step_str}' is infeasible "
                                 "due to exceeded bounds at initial conditions. "
                                 "If this step is part of a longer cycle, "
                                 "round brackets should be used to indicate this, "
@@ -894,15 +797,16 @@ class Simulation:
                                 "])"
                             )
                         else:
-                            this_cycle = self.experiment.operating_conditions_cycles[
-                                cycle_num - 1
-                            ]
+                            this_cycle = self.experiment.cycles[cycle_num - 1]
                             raise pybamm.SolverError(
                                 f"All steps in the cycle {this_cycle} are infeasible "
                                 "due to exceeded bounds at initial conditions."
                             )
                     cycle_sol = pybamm.make_cycle_solution(
-                        steps, esoh_solver=esoh_solver, save_this_cycle=save_this_cycle
+                        steps,
+                        esoh_solver=esoh_solver,
+                        save_this_cycle=save_this_cycle,
+                        inputs=user_inputs,
                     )
                     cycle_solution, cycle_sum_vars, cycle_first_state = cycle_sol
                     all_cycle_solutions.append(cycle_solution)
@@ -958,9 +862,9 @@ class Simulation:
 
         return self.solution
 
-    def run_padding_rest(self, kwargs, rest_time, step_solution):
-        model = self.op_conds_to_built_models["Rest for padding"]
-        solver = self.op_conds_to_built_solvers["Rest for padding"]
+    def run_padding_rest(self, kwargs, rest_time, step_solution, inputs):
+        model = self.steps_to_built_models["Rest for padding"]
+        solver = self.steps_to_built_solvers["Rest for padding"]
 
         # Make sure we take at least 2 timesteps. The period is hardcoded to 10
         # minutes,the user can always override it by adding a rest step
@@ -972,6 +876,7 @@ class Simulation:
             rest_time,
             t_eval=np.linspace(0, rest_time, npts),
             save=False,
+            inputs=inputs,
             **kwargs,
         )
 
@@ -984,6 +889,7 @@ class Simulation:
         t_eval=None,
         save=True,
         starting_solution=None,
+        inputs=None,
         **kwargs,
     ):
         """
@@ -1024,6 +930,7 @@ class Simulation:
             dt,
             t_eval=t_eval,
             save=save,
+            inputs=inputs,
             **kwargs,
         )
 
@@ -1169,8 +1076,8 @@ class Simulation:
         ):
             self._solver.integrator_specs = {}
 
-        if self.op_conds_to_built_solvers is not None:
-            for solver in self.op_conds_to_built_solvers.values():
+        if self.steps_to_built_solvers is not None:
+            for solver in self.steps_to_built_solvers.values():
                 if (
                     isinstance(solver, pybamm.CasadiSolver)
                     and solver.integrator_specs != {}